Update Project2018.py
Added comments throughout the scripts. Once the student was able to calculate the mean, min and max of the first column, it was then a simple task of writing scripts for the other columns. This was achieved by changing [:,0] to [:,1], whilst changing the definitions to reflect what the python script was executing. The use of the print commands allowed the calculations to be visually displayed.